chore: moved Chat type
