+++ frontend/src/types.ts
@@ -1,1 +1,5 @@
 export type Language = "javascript" | "java" | "python" | "go";
+export interface Chat {
+  from: string;
+  message: string;
+}
